Spark support: Add @turnkey/spark private package
