+++ examples/with-spark-schnorr/package.json
@@ -15,6 +15,7 @@
     "@scure/bip39": "^1.3.0",
     "@scure/btc-signer": "^1.5.0",
     "@turnkey/sdk-server": "workspace:*",
+    "@turnkey/spark": "workspace:*",
     "dotenv": "^16.0.3"
   },
   "devDependencies": {
