+++ packages/spark/README.md
@@ -1,0 +1,22 @@
+# @turnkey/spark
+
+[![npm](https://img.shields.io/npm/v/@turnkey/spark?color=%234C48FF)](https://www.npmjs.com/package/@turnkey/spark)
+
+[Turnkey](https://turnkey.com) Spark signer:
+
+If you need a lower-level, fully typed HTTP client for interacting with Turnkey API, check out [`@turnkey/http`](https://www.npmjs.com/package/@turnkey/http).
+
+API Docs: https://docs.turnkey.com/
+
+## Getting started
+
+```bash
+$ npm install @turnkey/spark
+```
+
+## Examples
+
+| Example                                            | Description                                                                                                                                                                                           |
+| -------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [`with-spark`](../../examples/with-spark/)         | Demonstrates [Spark](https://spark.money) operations using **Turnkey as the key custodian**.                                                                                                          |
+| [`with-spark-wdk`](../../examples/with-spark-wdk/) | Demonstrates [Spark](https://spark.money) operations through Tether's [`@tetherto/wdk-wallet-spark`](https://github.com/tetherto/wdk-wallet-spark) wallet API, with **Turnkey as the key custodian**. |
