+++ packages/spark/documents/docs.md
@@ -1,0 +1,5 @@
+---
+title: "Spark"
+description: "We have released a package that you can use to create a Turnkey-enabled Spark signer: [`@turnkey/spark`](https://www.npmjs.com/package/@turnkey/spark). See [here](https://github.com/tkhq/sdk/tree/main/examples/with-spark) and [here](https://github.com/tkhq/sdk/tree/main/examples/with-spark-wdk) for examples."
+mode: wide
+---
