+++ packages/spark/jest.config.js
@@ -1,0 +1,10 @@
+/** @type {import("@jest/types").Config.InitialOptions} */
+const config = {
+  transform: {
+    "\\.[jt]sx?$": "@turnkey/jest-config/transformer.js",
+  },
+  testPathIgnorePatterns: ["<rootDir>/dist/", "<rootDir>/node_modules/"],
+  setupFiles: ["dotenv/config"],
+};
+
+module.exports = config;
