+++ packages/spark/rollup.config.mjs
@@ -1,0 +1,3 @@
+import rollup from "../../rollup.config.base.mjs";
+
+export default (options) => rollup();
